+++ familyhelper-project-impl/src/main/java/com/dwarfeng/familyhelper/project/impl/bean/entity/HibernateUser.java
@@ -1,0 +1,61 @@
+package com.dwarfeng.familyhelper.project.impl.bean.entity;
+
+import com.dwarfeng.familyhelper.project.sdk.util.Constraints;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+
+import javax.persistence.*;
+import java.util.Optional;
+
+@Entity
+@IdClass(HibernateStringIdKey.class)
+@Table(name = "tbl_user")
+public class HibernateUser implements Bean {
+
+    private static final long serialVersionUID = -567003622114601946L;
+
+    // -----------------------------------------------------------主键-----------------------------------------------------------
+    @Id
+    @Column(name = "id", nullable = false, unique = true, length = Constraints.LENGTH_USER)
+    private String stringId;
+
+    // -----------------------------------------------------------主属性字段-----------------------------------------------------------
+    @Column(name = "remark", length = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    public HibernateUser() {
+    }
+
+    // -----------------------------------------------------------映射用属性区-----------------------------------------------------------
+    public HibernateStringIdKey getKey() {
+        return Optional.ofNullable(stringId).map(HibernateStringIdKey::new).orElse(null);
+    }
+
+    public void setKey(HibernateStringIdKey idKey) {
+        this.stringId = Optional.ofNullable(idKey).map(HibernateStringIdKey::getStringId).orElse(null);
+    }
+
+    // -----------------------------------------------------------常规属性区-----------------------------------------------------------
+    public String getStringId() {
+        return stringId;
+    }
+
+    public void setStringId(String stringId) {
+        this.stringId = stringId;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "(" +
+                "stringId = " + stringId + ", " +
+                "remark = " + remark + ")";
+    }
+}