+++ familyhelper-project-impl/src/test/java/com/dwarfeng/familyhelper/project/impl/configuration/CacheConfiguration.java
@@ -1,11 +1,15 @@
 package com.dwarfeng.familyhelper.project.impl.configuration;
 
 import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonProject;
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonUser;
 import com.dwarfeng.familyhelper.project.stack.bean.entity.Project;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.User;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
 import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.redis.formatter.LongIdStringKeyFormatter;
+import com.dwarfeng.subgrade.sdk.redis.formatter.StringIdStringKeyFormatter;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import org.dozer.Mapper;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -20,6 +24,8 @@ public class CacheConfiguration {
 
     @Value("${cache.prefix.entity.project}")
     private String projectPrefix;
+    @Value("${cache.prefix.entity.user}")
+    private String userPrefix;
 
     public CacheConfiguration(RedisTemplate<String, ?> template, Mapper mapper) {
         this.template = template;
@@ -33,6 +39,16 @@ public class CacheConfiguration {
                 (RedisTemplate<String, FastJsonProject>) template,
                 new LongIdStringKeyFormatter(projectPrefix),
                 new DozerBeanTransformer<>(Project.class, FastJsonProject.class, mapper)
+        );
+    }
+
+    @Bean
+    @SuppressWarnings("unchecked")
+    public RedisBatchBaseCache<StringIdKey, User, FastJsonUser> userRedisBatchBaseCache() {
+        return new RedisBatchBaseCache<>(
+                (RedisTemplate<String, FastJsonUser>) template,
+                new StringIdStringKeyFormatter(userPrefix),
+                new DozerBeanTransformer<>(User.class, FastJsonUser.class, mapper)
         );
     }
 }