+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - 建立实体以及维护服务，并通过单元测试。
   - com.dwarfeng.familyhelper.project.stack.bean.entity.Project。
+  - com.dwarfeng.familyhelper.project.stack.bean.entity.User。
 
 #### Bug修复